+++ services/api/src/app.ts
@@ -3,6 +3,7 @@ import { Elysia } from 'elysia'
 
 import { corsPlugin } from './plugins/cors'
 import { errorHandler } from './middlewares/error-handler'
+import { authModule } from './modules/auth'
 import { systemModule } from './modules/system'
 
 export const app = new Elysia()
@@ -13,6 +14,6 @@ export const app = new Elysia()
   )
   .use(corsPlugin)
   .use(errorHandler)
-  .group('/api', (api) => api.use(systemModule))
+  .group('/api', (api) => api.use(systemModule).use(authModule))
 
 export type App = typeof app